requirements.txt
whisper
tqdm
psutil
pyannote.audio
torch
torchaudio
ffmpeg-python